requirements.txt
pydantic==2.12.4
kafka==1.3.5
PyYAML==6.0.3
requests==2.32.5
ratelimit==2.2.1
pydantic-settings==2.12.0